Add initial tests for SetClip_PathFill
Debugging for imaging issue in [1] revealed unexpected bad behavior
coming out of Windows GDI API.  Microsoft documentation [2] indicates
that if a bad clip is provided (e.g., cause an error with
IntersectClipRect()) then the current clipping region should be
unaffected. The disappearance of images has shown that this isn't
necessarily true in all cases, and is something to be avoided.

Update CGdiDeviceDriver::SetClip_PathFill() to watch for an error
condition from IntersectClipRect() and return failure if it happens.

Add some new tests for this including to illustrate triggering this
GDI condition.

[1] https://crbug.com/1019026
[2] https://docs.microsoft.com/en-us/windows/win32/api/wingdi/nf-wingdi-intersectcliprect

Bug: chromium:1019026
Change-Id: I336617399bf90b511de5d6fbecf3da829fbe4de6