Use PDF page base clip for Win32 rendering
Introduce a base clip for a PDF page.  This base clip is the same as the
initial clip rectangle that is created when rendering for a page is
started; it is now also remembered for later use.

Utilize this base clip for rendering with the Win32 device.  This
provides extra protection in some cases against otherwise valid clip
paths for which the coordinates are so large that they trigger errors
when used with the Windows GDI API.

It is not easy to apply this in all cases since some paths could be
impacted by clipping too early (e.g., if applying it affected fill
mode for complicated paths), and such paths are provided straight to
GDI.  Start with an initial easy fix of applying this clip against
rectangular paths, which resolves an issue seen from at least one PDF
generated by Microsoft Word that had a rectangular clip extending
hundreds of millions of pixels beyond the page boundaries, having no
effective use yet was triggering coordinate issues and causing images to
be erroneously discarded.

Bug: chromium:1019026
Change-Id: I2c5e6e1cd6444b8d4357bd812453d89eabf1ab81