Add a test for a non-embedded font with illegal CIDToGIDMap string.
The PDF Reference, Sixth Edition, specifies that a CIDToGIDMap "may
appear only in a Type 2 CIDFont whose associated True Type font
program is embedded in the PDF file" (Section 5.6, pp 437). But some
PDFs, which use Type 2 CIDFont, contain CIDToGIDMap as string
"Identity" even when the font is not embedded.

This CL adds a test case with the incorrect rendering suppressed.
PDFium is rendering CID <0004> as unicode <0041> ('A') by looking up
ToUnicode mapping and ignoring the illegal CIDToGIDMap, while the
expected result should be '!' (glyph index = 4) because it follows
the rule that when CIDToGIDMap's value is "Identity", glyph index
should be the same as charcode.

Bug: chromium:1021762
Change-Id: If9dd2cb8ff03142a39ca86076e72f75f281ec747