Add another malloc(0) test.
In the past, we've pondered letting malloc(0) return a non-null
zero-page (or otherwise umappable) pointer to immediately catch
any attempt to reference it. The simplest implementation for
this would return the same pointer for all malloc(0) calls, which
would then make it impossible to tell if two objects actually shared
the same malloc(0)'d buffer. It is doubtful that PDFium actually
depends on this behavior, but add a test to enforce this should
someone down the road want to implement the malloc(0) protection
(returning distinct unreferenceable pointers would be legitimate).

Change-Id: I6f1fd46dc9e2560ab1efc475dcd33ba85f7c11d3