[FXCRT] Add unit tests for CFX_FloatRect::GetInnerRect()
There were none before, now there can be some.  This exposes an issue at
the edges of the integer range, where a float larger than the maximum
value of a signed 32 bit integer returns a value of the maximum negative
integer.

Similar to the unexpected behavior uncovered during debugging of Windows
print imaging issue in crbug.com/1019026.  This is the GetInnerRect()
equivalent to the change in [1].

[1] https://pdfium-review.googlesource.com/c/pdfium/+/62130

Change-Id: Ic711dc3d8ee3a52098e6dfb07f6ee4aaf364938c