Ensure CFWL_Widgets are not locked during destruction.
This is a strong signal that there is active code below us on
the stack that requires the widget to persist, yet through a twisty
series of callbacks we may reach this spot. Ideally, we'd notice
this earlier and unwind cleanly, but for now it is likely that the
only documents that would reach this case are trying to muck with us,
so aborting may be appropriate.

Bug: chromium:1032422
Change-Id: I78e242cbc157c38666ffb1f248edcde2d79dbca3