Save encrypted files without IDs in CPDFSecurityHandlerEmbedderTests.
Add methods to CPDF_Parser and CPDF_CrossRefTable to strip the /ID entry
from the PDF trailer. Use this in CPDFSecurityHandlerEmbedderTests to
exercise the PDF creation code paths where they attempt to write out new
encryption dictionaries. This uncovered at least 2 bugs in the code, so
some of the new test code is disabled for now, with TODOs to fix them.

Bug: chromium:1032090,pdfium:1440
Change-Id: Ic285e4ed7ff89b3dc9c604f3e51c5345a4f64f27