Fix password encoding when generating PDFs.
Per https://crbug.com/pdfium/1194, PDFium accepts passwords in either
Latin1 or UTF8 encoding, and will re-encode, if necessary, to correctly
decrypt PDFs. When the encrypted PDFs are copied and saved, the copy
needs to be encrypted with the password with the proper encoding.

Do so by remembering the re-encode status in CPDF_SecurityHandler. Then
use that knowledge to encode passwords correctly in CPDF_Creator.
Re-enable CPDFSecurityHandlerEmbedderTest code that was disabled due to
this issue, and add new TODOs for test cases that still do not work due
to other issues.

Bug: pdfium:1440,pdfium:1441
Change-Id: I4fe27403db06a1a5dcb90d307fe00af181a1ad22