Avoid crash in CPDF_SecurityHandler::GetUserPassword().
For r2 and r3 encrypted PDFs, the value for the /O key is normally a
32-byte string. CPDF_SecurityHandler::GetUserPassword() recently started
using spans, and that caught an OOB error for /O keys shorter than
32-bytes. While a hard crash is better than an OOB read, gracefully
handling the situation is best. Change GetUserPassword() to do that.

For testing, add hand-edited copies of encrypted_hello_world_r2.pdf and
encrypted_hello_world_r3.pdf, with 31-byte /O keys, and the xref offset
adjusted. Then make sure embedder tests that try to load these bad PDFs
do not crash.

Bug: pdfium:1436
Change-Id: I873eead7c1dc45c9c4bbd41044f8fa5b77d7f558