Avoid using std::function to improve performance.
In CFX_ImageTransformer, stop using std::function in places that just
need a lambda. Use a template optimization instead, to avoid
std::function which has some overhead. In loop heavy code like
CFX_ImageTransformer, this improves performance by up to 10% in release
builds.

Also shuffle CFX_ImageTransformer code so that the templates are not
exposed in the header file.

Change-Id: Ia833843ea4768eef82ce1986993cb1f1e86a8895