Only hash pixel data in EmbedderTest::HashBitmap().
Add a new EmbedderTest::HashBitmap() implementation that only hashes the
pixels in a given bitmap, and ignores the data in the gaps at the end of
rows. Per prior discussion on [1], the data in the gaps can in theory be
uninitialized.

Rename the existing implementation to HashBitmapDeprecated(), and switch
over all HashBitmap() callers except EmbedderTest::CompareBitmap().

[1] https://pdfium-review.googlesource.com/64290

Change-Id: I24bd143f6b33c7ef86e26dd09d1cb874dea1675f