Ensure RetainPtr<T>(p)->Method() stays alive for duration of Method
Even if Method() removes the last reference. A modest proposasl that is
nonetheless mind-crushingly slow I would expect.

C++ allows operator->() to return either the raw type, or another object
by value which has an operator->() that returns the raw type. So now
when we write

  m_pObj->Method();

we first get a temporary |Local| constructed that increment the ref
during its lifetime, and then returns the raw type. so the count
remains in effect until the next ; where C++ cleans up all these
invisible temporaries.

Unfortunately,

  x = m_pObj->width;

also incurs this cost, but maybe the optimizer can realize that
it need not be done.

The idea is we shouldn't need

  RetainPtr<T> protect(this);

in various methods to keep things safe.

Because there are times when we need to actually de-ref without
incrementing, make operator*() behave differently than operator->().

Change-Id: I1ea9a13d8e46383061e532934b134ee4d1b3f8c7