CPDF_RenderOptions::bForceDownSample is always false.
So remove it. In turn, FXDIB_ResampleOptions::bInterpolateDownsample is
now always false, so remove it. In turn, StartQuickStrech() and
ContinueQuickStretch() are now unused, so remove them. In turn,
m_bFlipX, m_bFlipY, and m_LineIndex are now unused, so remove them.

It is just as well since forced downsampling without any interpolation
always looks terrible ...

Change-Id: I798bec7494d0ffa0744a80b220dc29276b2de220