CPDF_ImageLoader::Start() is always passed a cache
Because it is only called in one place using CPDF_RenderContext's
page cache. CPDF_RenderContext's page cache is not null because it
is always created from either a page (and everywhere we create a page
we add a cache on the next line), or from another RenderContext passed
through DrawPatternBitmap(), which comes from
CPDF_RenderContext::GetPageCache().

-- In turn, this makes GetPageResource() unused.

Change-Id: Ic5b544c80e6d93cd379b3c5ed956200e8d6fd811