Rename pixel test bug_1021762 to bug_1457.
The original pixel test bug_1021762.in doesn't reflect the original
issue from crbug.com/1021762, but it still reflects a valid rendering
issue. This CL renames bug_1021762.in to bug_1457.in and updated the
expected result and SUPPRESSIONS list accordingly. Another pixel test
will be uploaded for crbug.com/1021762 in a separate CL.

Bug: pdfium:1457
Change-Id: I31b2ba040626644810b96ac6b34c8b40e2f86f6d