Move CPDFXFA_Context creation to FPDFDOC_InitFormFillEnvironment()
Currently the CPDFXFA_Context creation occurs in LoadDocumentImpl()
regardless of whether --disable-xfa has been passed to pdfium_test.
We want to defer the context creation to
FPDFDOC_InitFormFillEnvironment() and make it conditional to whether the
runtime flag has been passed.

Add a new field |xfa_disabled| to version 2 of _FPDF_FORMFILLINFO.
|xfa_disabled| was selected over |xfa_enabled| to allow for backwards
compatibility, as |xfa_disabled| defaults to false/0 upon
initialization.

This CL changes when the embedder gets a callback for when XFA is not
supported.

Fixed: pdfium:1433
Change-Id: I01da4fd75127815728d2c4dae357e4f22447a6c9
Bug: 1433