Rename CFWL_Widget::{Get,Set}FFWidget() to {Get,Set}AdapterIface().
Although a CXFA_FFWidget is a CFWL_Widget::AdapaterIface, this interface
exists so that the fwl/ layer has no knowledget of FF widgets, and hence
the name shouldn't contain that term, either.  As a result, it becomes
clearer why the result of one of these calls needs to be cast to a
CXFA_FFWidget by the upper layers, as applying such a cast to a call
named GetFFWidget() was unusual looking.

Change-Id: Ifefc243503b04ff17105b40931cd12be0e7a1fc4