Update patch set 8
Patch Set 8:

Can you respond to the comments from patch set 4? e.g. Mark the comments as "done" if you took the suggestion in the comments. Then hit "reply" at the top to send out the comments.

Between patch set 4 and 8, the code changed to modify the scoring in GetSimilarValue(). Can you discuss why this is better than the earlier patch set?

Patch-set: 8