Reformat PDFium GN files.
`gn format` recently changed its formatting behavior for deps, source,
and a few other elements when they are assigned (with =) single-element
lists to be consistent with the formatting of updates (with +=) with
single-element.

Now that we've rolled in a GN binary with the change, reformat all files
so that people don't get presubmit warnings due to this.

Bug: chromium:1041419
Tbr: thakis@chromium.org
Change-Id: I879c199419bdd2a9fe09ab09d1a0ce6e178b4fee