Avoid redundant EnumFontList() invocations.
Instead, use the (idempotent) CFX_FontMapper::LoadInstalledFonts()
which makes this exact call but at most once per font mapper. Continue
to call this in CreateFontStream() even though it appears redundant
because I've not run down all possible call paths.

For the simple_xfa.pdf document, this results in a 4x reduction in
the number of times the font directories are scanned. This is a
contributor to the referenced bug, but other problems still remain.

Bug: pdfium:blah
Change-Id: Ie0329dd329e586db02436b963352a66a6c4e751e