Add a pixel test for non-embedded Symbol font.
This pixel test does not embed any fonts, but tries to use the
"Symbol,Bold" TrueType font to render 'a' as an alpha symbol. Other
PDF viewers does the correct font substitution and renders an alpha
symbol. Illustrate the issue where PDFium just renders a bold 'a'.
The expected output is incorrect, but is as is to make the test pass.

Change-Id: I3df9f2e8a98944ec8d06fc591757facffaf88e27