Fix the rendering of scientific symbols correctly
The file in the bug contains text which has to be rendered
in "Symbol" font. This font is not embedded in the pdf file.
When rendering call comes for text that corresponds to Symbol font, the
font is fetched from the system. Even though symbol font is present on
the system, it is not picked up.

Currently, the font is fetched by following the two steps:
1. Find whether there is a matching font in the installed fonts on
   the system. This call returns font face name if found in the list of
   installed fonts.
2. Getting the font face for the required font.
   This call again iterates over all the installed fonts
   and returns the font face for the best matching font.

For the case in question:
Incorrect font whose name is a prefix of desired font is being picked up
due to GetSimilarValue() yielding the same result.

To fix this, check if the family name matches the font name exactly,
then increase | iSimilarValue | by 16.

Bug: pdfium:1449
Change-Id: I23d5c3f5e4f26981fc6a9af2ccd0024c3f9b6102