Remove the prefix in BaseFont when trying to find a substitute font.
According to PDF reference 1.7 section 9.6.4, for a font subset, its
BaseFont will begin with a 6-letter arbitrary prefix followed by a plus
sign(+), for example "TAXXCX+Arial" is a subset font of Arial family.

To find a better substitute font, we need to correctly extract the font
family name instead of trying to find a match for the arbitrary prefix.

This CL adds the step to remove the possible prefix in BaseFont in
CFX_FontMapper::FindSubstFont() and improves the rendering result.

Bug: chromium:1021762
Change-Id: Id3f88f83202fbf9811a2c4863c08d9776455caa8