Remove CFFL_FormFiller::WndtoPWL() and FFLtoWnd()
There are really only two relevant coordinate systems in this code,
FFL as used by code in formfiller/, and PWL as used by code in pwl/.
The Wnd system, as contained in the names above, is really just the
FFL system. Hence "WndtoPWL" is really just "FFLtoPWL", and "FFLToWnd"
is really just "FFLtoFFL" and a no-op passthrough.

Change-Id: I559e7d4e8008fa1504f1cbb21ef1b6382b59f0a5