Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> 1. please strongly consider switching to go/lucicfg instead of manually crafting these protobufs (chromium/src did this and they were super happy).

The documentation looks long. I'll make a note of it and try it out next time. This CL follows the same strategy as the one I used in https://crbug.com/pdfium/1340, so it's well known to me.

> 2. I have no rights to lgtm here && no expertise in your configs, perhaps choose a different reviewer?

PDFium Mac bots use some amount of macOS 10.13 machines from luci.flex.ci right now. The questions for you as a reviewer are:

a) Is the "mac_mojave" buider_mixins with s/13/14/ the right way to specify 10.14 machines? I would assume, so, but figured it's good to sanity check.
b) If we switched our entire PDFium build/trybot workload from 10.13 to 10.14, is that ok?

I'll definitely get someone with the +1 rights to approve afterwards.

Patch-set: 1