Remove some dead code from CFX_CSSSyntaxParser
The kUnknownRule case can never be hit because there are no paths
that set m_eMode to kUnknownRule. Additionally, a stray close brace
while looking for a ruleset is always an error since we don't support
recursive braces anyways (such as might have been introduced with
at-rules or the like).


Change-Id: I52d89099bf9bc4d6559c1241ada507dc466ce3a5