CFX_TxtBreak::EndBreak_SplitLine() always returns false.
The change at https://pdfium-review.googlesource.com/c/pdfium/+/2938
removed unreachable code, and along with it, the possibility of
a |true| return from the function. Make it a void instead.

Also remove an unused CFX_BreakPiece local (it's a POD struct apart
from an vector and a retain ptr which remain empty and null
respectively after default construction).

Change-Id: I38474ec1b8979a285c70be2bac70a0b271bd9a64