Update patch set 1
Patch Set 1:

thakis: Mind taking a look while tsepez is OOO? This is basically the same as https://pdfium-review.googlesource.com/c/pdfium/+/66213, because we decided to target 10.15 instead of 10.14. Both this and that CL will be reverted when we verify we can switch PDFiumn to 10.15.

Patch-set: 1
Reviewer: Gerrit User 5475 <5475@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false