Remove the prefix in BaseFont when trying to find a substitute font.
According to PDF 32000-1:2008 spec section 9.6.4, for a font subset, its
BaseFont will begin with an arbitrary prefix of 6 uppercase letters
followed by a plus sign(+), for example "YATXCX+Arial" is a subset font
of Arial family.

To find a better substitute font, we need to correctly extract the font
family name instead of trying to find a match for the arbitrary prefix.

This CL adds the step to remove the possible prefix in BaseFont in
CFX_FontMapper::FindSubstFont() and improves the rendering result.

Bug: pdfium:1475
Change-Id: Id3f88f83202fbf9811a2c4863c08d9776455caa8