Add FXSYS_IsUpperASCII().
Add FXSYS_IsUpperASCII() to check whether a character is an uppercase
ASCII letter.

This is a follow-up CL for
https://pdfium-review.googlesource.com/65752/. This new function helps
limit IsStrUpper() to only check for uppercase ASCII letters instead of
all uppercase Unicode characters.

This CL adds more callers for FXSYS_IsUpperASCII() to replace all
uppercase ASCII character checks across PDFium and adds a matching unit
test for the function.

Bug: pdfium:1475
Change-Id: I67d12acebcbe3c595fb888ae8fbf6c08432b80ed