Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (15 comments)
> 
> > Patch Set 1:
> > 
> > (10 comments)
> > 
> > Can we add embedder tests for the same?
> 
> We are using core Apis to test this( e.g. cpdf_array,cpdf_annotcontext from core/fpdfapi/page). Using core apis is not recommended to use in embedder tests. That's the reason test is written in fpdf_annot_unittest.cpp.
> 
> Resolved all other review comments.

From what i have seen we should have a getter and setter API's. This would ensure completeness and the ability to an embedder to exercise the same if and when needed. I would expect the getter to be either in this CL or checked in before this CL so that getters are used to test the setters.

Please queue dry run for the CL.

Patch-set: 3
Reviewer: Gerrit User 7659 <7659@fca2d64a-0420-3532-b200-12848c95a47b>