Add a pixel test for /SMaskInData in JPEG2000 images.
In images using the JPXDecode filter, if /SMaskInData is set to 1, then
the image data stream includes soft-mask values for use as a source of
mask shape or mask opacity.

The sample JPEG2000 image is generated from a PNG with alpha using
OpenJPEG. The image has to be a binary for /SMaskInData 1 to work. i.e.
No additional filters.

Bug: pdfium:1469
Change-Id: Id40ec19ee587fcdfd930e1f3a6644326990948a3