Support custom trailers in fixup_pdf_template.py
An existing challenge is checking in PDFs with custom trailers with
an accompanying .in file. Add a {{size}} token handle in the script to
maintain some automation in the trailer generation.

{{trailer}} will still generate the generic trailer:
trailer <<
  /Root 1 0 R
  /Size N
>>

With {{size}}, input files can have custom trailers like:
trailer <<
  /Info 2 0 R
  /Prev 3 0 R
  /Root 4 0 R
  {{size}}
>>

At the same time, the developer will not have to worry about counting
the number of objects in the PDF.

Change-Id: If86d4ab2069c21d609f3241b5f375bb8d135924d