Fixed rendering issue for PDFs with incorrect first-page hint data.
For a linearized PDF, sometimes the hint data can be incorrect. For
example, The test file in crbug.com/1012237 has hint data which points
the first page to the wrong object number, which puts a wrong but
existing object number into |m_PageList[0]|. This results in
GetPageDictionary() returns a dictionary which doesn't belongs to a
page-type object.

To fix the issue, this CL adds an extra check for the object type to
be "Page" when GetPageDictionary() is trying to get dictionary of a
single loaded page by page number. If the type doesn't match, mark the
page as unloaded and keep going through the Page tree to find the
correct object and return its dictionary.

Since the original GetPageDictionary() doesn't check object type and
it has been used in several unit tests which use mocked-up incomplete
page structure, we need a function that's doesn't enforce the page
type check for testing purpose as well. To solve this issue, this CL
adds a new function, GetPageDictionaryWithTypeCheckOption(), which has
basically the same functionality as GetPageDictionary() but with an
extra parameter |check_type_is_page| indicating whether page type check
is enforced. So that the unit tests can call the new function instead
with parameter |check_type_is_page| = false. And the new
GetPageDictionary(), which is used in the actual rendering process,
becomes a caller for this new function with parameter
|check_type_is_page| = true.

Bug: chromium:1012237
Change-Id: I1e8745a48ed0cabe5478e9e8e1dda51fa0a8d53f