Fix rendering issue for PDFs with incorrect first-page hint data.
For a linearized PDF, sometimes the hint table can contain incorrect
data. For example, The test file in crbug.com/1012237 has hint table
which points the first page to the wrong object number, and set this
wrong object number into |m_PageList[0]|. This results in
GetPageDictionary() fetches a dictionary from an existing but
non-page-type object for the first page and renders it blank.

To fix the issue, this CL creates CPDF_Document::IsObjValidPage() to
validate if an object number points to a valid "Page" type object, and
add this validation in CPDF_Document::LoadPages(). If the check fails,
don't set this wrong object number to |m_PageList[0]|. Instead, traverse
the page tree to update |m_PageList|.

Bug: chromium:1012237
Change-Id: I1e8745a48ed0cabe5478e9e8e1dda51fa0a8d53f