Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> This CL doesn't contain a test due to the complexity of creating a PDF with "incorrect" linearized data. Chatted offline with Lei about it and he suggest ask for permission from the reporter to check in the sample PDF they provided in the ticket.

Haven't reviewed yet, but quick comment: The sample PDF seems to contain a lot of PII I wouldn't want checked in, personally. (Git is forever.) If we can't get permission, or it just seems like a bad idea generally, is it possible to start with a correct linearized PDF, and then hack it somehow to be broken in the right way?

Patch-set: 5
Reviewer: Gerrit User 7613 <7613@fca2d64a-0420-3532-b200-12848c95a47b>