Add a pixel test for page index clobbering
If given an object number that is not for a /Page, CPDF_Document::GetPageIndex()
can cache the incorrect object number for a page index, leading to problems such
as drawing blank pages.

The pixel test uses a /GoTo action to a /Pages node trigger this problem,
drawing a blank page instead of the expected gray rectangle.

Bug: pdfium:1491
Change-Id: I9dc5a8fe401f81262097dd6bf0b51229a386d1ac