Update pixel/corpus tests to handle expected results for skia.
With Skia/SkiaPaths enabled, the rendering results can be correct but
slightly different compared to AGG's results. This CL allows
run_pixel_tests.py and run_corpus_tests.py to also compared to an
acceptable rendering results generated by Skia/SkiaPaths, so that if the rendering
result is not completely the same as AGG's expectation, it won't cause
failure of pixel/corpus test.

This CL also provides skia-based and skia-platform-based expected
results for testing the script change.

Bug: pdfium:1501
Change-Id: I237c70bdec2640995f9cb8602d0b32addc0598e0