XFA: Test resolving variables cross-variables context
Attempt to cover a portion of CFXJSE_Engine::QueryVariablesValue() by
introducing variables in different contexts whose names shadow each
other. We should be able to refer to each set via a specific script
node associated with each context.

The test unfortunately is based on trying to cover program logic
rather than a specification of correct behaviour. In particular,
requiring a script to be parented by a variables context (script3
in the test) is questionable.

Change-Id: I098845e75442e93e26fbf7f954a0a0d18e9705ed