Fix `FindPageIndex` search algorithm.
Regression introduced in 5b4135d22ddd1450327931ee04fb3f14ada95471.

Required when reaching a leaf node of the Pages tree during recursion of
`FindPageIndex`.

A check for 0 `skip_count` is required before decrementing. `skip_count`
is unsigned so whenever this is decremented, the check against the
"Count" key of the "Pages" dictionary always succeeds and `skip_count`
is incorrectly reduced again.

Change-Id: If82b77ab6e1ba2ad44b8d138a7ebffecb079313a