Make PRESUBMIT.py accept SkiaPaths expected results.
Using Skia and Skiapaths sometimes lead to different rendering results.
For a pixel/corpus test case, if both rendering results are acceptable,
a Skiapaths specific expected result is needed for this test besides an
skia expected result (different from the AGG expected result and the
name is in the form of
"NAME_expected_skia(_(win|mac|linux))?.pdf.#.png").

This CL modifies PRESUBMIT.py to accept Skiapaths specific expected
results with names in the form of
"NAME_expected_skiapaths(_(win|mac|linux))?.pdf.#.png", and added a
test for _CheckPNGFormat().

Bug: pdfium:1501
Change-Id: Ie8def15b9a18e86445e49620567d54b193d6bc0f