Update pixel/corpus tests to handle expected results for Skia.
With pdf_use_skia enabled, the rendering results can be correct but
slightly different compared to AGG's results. This CL allows
run_pixel_tests.py and run_corpus_tests.py to also compared to
acceptable rendering results generated by Skia, named in the format of
"NAME_expected_skia(_(win|mac|linux))?.pdf.#.png". So that if the
rendering result of a PDF is not completely the same as its AGG's
expectation, as long as it finds a match with a Skia expected rendering
result, it won't cause a failure in pixel/corpus test. This CL also
provides skia-based and skia-platform-based expected results for
testing the script change.

In case that SkiaPaths, Skia and AGG all generate correct results and
these results are different from each other, we will need to add
expected results for SkiaPaths in a new format of
"NAME_expected_skiapaths(_(win|mac|linux))?.pdf.#.png", besides the
expected results of Skia and AGG.

Bug: pdfium:1501
Change-Id: I237c70bdec2640995f9cb8602d0b32addc0598e0