Merge and extend FocusableAnnotSubtypes tests.
Instead of testing the 3 related APIs separately, test them at the same
time as they can naturally be used together.

Also add some FORM_SetFocusedAnnot() calls to see the effect of calling
FPDFAnnot_SetFocusableSubtypes(). This also reveals an issue with
FORM_SetFocusedAnnot() when it comes to focusing popups and widgets.

Bug: pdfium:1507
Change-Id: Ic8ee8bb897b0f92f02c6c0b55f286f735ee84cd3