Remove spaces adjacent to bracket characters in some PDFs
Currently, brackets in most PDFs are styled as such:
[ (item1) (item2) (item3) ]

Remove the spaces adjacent to brackets to style as such:
[(item1) (item2) (item3)]

The new style is more compact and easier on the eyes for applications
like nested lists. Apply the style on rectangles.in and hello_world.in,
as those are the two files developers are most likely to look at when
styling new files.

Change-Id: Ic6144caf0369754bfb40acff7a5afdb4c6e30725