Update pixel/corpus tests to handle expected results for Skia.
With pdf_use_skia enabled, the rendering results can be correct but
slightly different compared to AGG's results. This CL allows
run_pixel_tests.py and run_corpus_tests.py to also compared to
acceptable rendering results generated by Skia, named in the format of
"NAME_expected_skia(_(win|mac|linux))?.pdf.#.png". So that if the
rendering result of a PDF is not completely the same as its AGG's
expectation, as long as it finds a match with a Skia expected rendering
result, it won't cause a failure in pixel/corpus test.

This CL also provides PNG files of Skia-based and Skia-platform-based
expected results for testing the script change:

- For bug_842.in, it contains a gradient transition that Skia uses a
  different algorithm to render compared to AGG. Therefore, upload its
  Skia rendering results for testing script change.

- For bug_733528.in, Skia and AGG both render it with the correct font
  face, but due to Skia does matrix and font size scaling differently
  in the process (scaling happens later in the process without
  rounding numbers of glyph positions, sometimes uses fast computing
  methods) compared to AGG, the glyphs end up with slightly different
  rectangle boundaries. Since this test case involves font rendering,
  which will trigger different behaviors among different platforms, it
  will serve the purpose of testing the script change on
  Skia-platform-based expected results.

In case that SkiaPaths, Skia and AGG all generate correct results and
these results are different from each other, we will need to add
expected results for SkiaPaths in a new format of
"NAME_expected_skiapaths(_(win|mac|linux))?.pdf.#.png", besides the
expected results of Skia and AGG. Currently we don't have an existing
test case that can be rendered differently and correctly among using
all three rendering methods. Therefore there is no PNG file uploaded for
testing this SkiaPaths specific change.

Bug: pdfium:1501
Change-Id: I237c70bdec2640995f9cb8602d0b32addc0598e0