Make UnownedPtr<> interoperate more transparently with raw pointers.
There is now interest in using some sort of unowned ptr variant
beyond PDFium, but concerns exist about minimizing the amount of
code that needs to change during the adoption of it. Demonstrate
the effects in PDFium of adopting such a new interface:

-- Add an implicit operator T*() which is preferred to .Get().
-- The existing operator=() will become preferred over .Reset().
-- Then remove comparison operators which become ambiguous once
   demotion to raw pointers is implicit.

Also update the node iterator template, because it is instantiated
against both RetainPtr and UnownedPtr. This is possible because
both of these have the same "shaped" API.  In turn, update RetainPtr
to match the new shape:

-- Add an explicit operator T*() (not implicit, want to limit use)
-- Allow assignment from the underlying type in place of .Reset().

Change-Id: I50828f50b3debb0b1b9ddd494cb640048933d92b