Rewrite and simplify CPDF_FormField::IsItemSelected()
The PDF spec indicates that the /V entry of a choice field should take
precendence over the /I entry for identifying the selected options.
Additionally, the /I entry was only introduced in PDF 1.4, meaning that
there can exist PDF choice fields that contain an /V entry and no /I
entry, which is not correctly handled by the current implementation of
the method.

Bug: pdfium:1505
Change-Id: I7646e2f6d69e581ea68c55da11a24b26e26ee45d