Fix embedder test FPDFFormFillEmbedderTest.BUG_1281.
Test FPDFFormFillEmbedderTest.BUG_1281 fails for Skia and SkiaPaths
because |bRgbByteOrder| is ignored when setting color type for bitmap
inside CFX_SkiaDeviceDriver, while it's supposed to enforce the reverse
byte order of color.

For color rendering, Skia and SkiaPaths are always using color type
|kN32_SkColorType|, of which the value changes between
|kRGBA_8888_SkColorType| (byte order is R,G,B,A) and
|kBGRA_8888_SkColorType| (byte order is B,G,R,A) depending on whether
the program is running on a big endian processor or not.

To make the byte order of color changeable instead of depending on the
processor, this CL makes Skia and SkiaPaths both respect the value of
|bRgbByteOrder| that's passed into CFX_SkiaDeviceDriver, and use
|m_bRgbByteOrder| to hold this value and keep the color type consistent
throughout the rendering process just like AGG.

In addition to the fix, this CL also adds the expected hash values for
Skia/SkiaPaths rendering results in the embedder test, since the small
discrepancies for rendering small rectangles are expected with
Skia/SkiaPaths enabled.

Bug: pdfium:1500
Change-Id: Iad0762a04cc3f266accb4d48177ccb8fd95af894